requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.17.0
google-auth==1.14.3
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-resource-manager==0.30.1
google-cloud-secret-manager==1.0.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.29.0
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
